Move amp_state definition to include/hamlib/amp_state.h
